+++ src/main/java/com/github/jlangch/venice/impl/specialforms/SpecialForms.java
@@ -79,7 +79,6 @@ public class SpecialForms {
 					"defmacro",
 					"macroexpand",
 					"ns",
-					"namespace",
 					"ns-unmap",
 					"ns-remove",
 					"ns-list",